Fix nil params when transitioning from a modal to another (#30)
As Iván pointed out, when opening a modal from another modal, we get an error because params are null for a brief moment. This will fix it, I hope.

This PR refactors the entire logic and encapsulates the whole modal state into a single object to force batch all of the different setStates, which were causing inconsistent states between modal renders, which caused null errors on the params object.