Update dev deps not part of build and actions (#35)
Updates all devs deps not part of the build process, meaning everything except `rollup` stuff as it as a lot of breaking changes. I'll take a look later!

Also updates the Github actions to the latest major versions available.
AND changes the NodeJS matrix to use node `14` and `18` as well.